refactor(export): delete the hand-rolled PDF writer
Both PDF paths (table + records) now render through the designed gopdf
renderer, orphaning the raw-PDF-syntax writer: the WinAnsi text encoder,
the transliteration fallback (the source of Nguyễn -> Nguyen mangling),
the manual xref assembly, and the old record pager. pdfColumnWidths is
the one survivor — it moves to pdf_design.go.

TestTransliterateRune goes with the code it locked (that behaviour is
exactly what the migration removes); the old multi-page helpers test is
replaced by a rendered-page-count smoke test.
